+++ package.json
@@ -4,6 +4,7 @@
     "classnames": "^2.2.6",
     "core-js": "^2.5.5",
     "fixed-data-table-2": "^1.0.0-beta.5",
+    "isomorphic-fetch": "^2.2.1",
     "preact": "^8.4.2",
     "preact-compat": "^3.18.4",
     "react-aria-modal": "^3.0.0",
@@ -13,7 +14,8 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0"
+    "redux-thunk": "^2.2.0",
+    "webpack-manifest-plugin": "^2.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
@@ -27,7 +29,7 @@
     "aws-sdk": "^2.205.0",
     "babel-core": "^7.0.0-bridge",
     "babel-loader": "^8.0.0",
-    "css-loader": "^1.0.0",
+    "css-loader": "~1.0.0",
     "ejs-webpack-loader": "^2.2.2",
     "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.1.0",
@@ -54,7 +56,8 @@
     "url-loader": "^1.0.1",
     "webpack": "^4.27.1",
     "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.0"
+    "webpack-dev-server": "^3.1.0",
+    "webpack-merge": "^4.1.5"
   },
   "license": "MIT",
   "name": "cabinet",
@@ -65,7 +68,8 @@
   "scripts": {
     "awsUtils": "node aws/utils",
     "lint": "eslint . && stylelint src/stylesheets/*.css",
-    "start": "webpack-dev-server"
+    "start": "webpack-dev-server --config webpack.browser.config.js",
+    "build": "NODE_ENV=production webpack --config webpack.browser.config.js && NODE_ENV=production webpack --config webpack.server.config.js"
   },
   "version": "0.0.4",
   "browserslist": [
