+++ src/index.html.ejs
@@ -12,7 +12,7 @@
   <meta property="og:image" content="https://c1.staticflickr.com/6/5564/14726222014_6dc5cc1cdb_b.jpg">
   <title>CabinetVotes</title>
   <% typeof css !== 'undefined' && css.forEach(function(file) { %>
-    <link href="/<%- file %>" rel="stylesheet preload" as="style">
+    <link href="<%- file %>" rel="stylesheet preload" as="style">
   <% }); %>
 </head>
 <body>
@@ -22,7 +22,7 @@
     <% } %>
   </div>
   <% typeof js !== 'undefined' && js.forEach(function(file) { %>
-    <script src="/<%- file %>" defer></script>
+    <script src="<%- file %>" defer></script>
   <% }); %>
   <script async="true" charset="utf8" src="https://platform.twitter.com/widgets.js"></script>
 </body>
