+++ src/polyfills.js
@@ -1,3 +1,4 @@
 import 'core-js/es6/map';
 import 'core-js/es6/set';
 import 'core-js/es6/promise';
+import 'isomorphic-fetch';
